refactors board acceptance tests
